img2excel/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "img2excel"
version = "1.0.0"
description = "将任意图片转换为Excel像素画的Python包"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Charlesshen2", email = "463674542@qq.com"}
]
maintainers = [
    {name = "Charlesshen2", email = "463674542@qq.com"}
]
keywords = ["image", "excel", "pixel", "art", "conversion", "graphics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Office/Business :: Financial :: Spreadsheet",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.7"
dependencies = [
    "Pillow>=8.0.0",
    "openpyxl>=3.0.0",
]

[project.scripts]
img2excel = "img2excel.cli:main"

[project.urls]
Homepage = "https://github.com/Charlesshen2/img2excel"
Documentation = "https://github.com/Charlesshen2/img2excel#readme"
Repository = "https://github.com/Charlesshen2/img2excel"
"Bug Tracker" = "https://github.com/Charlesshen2/img2excel/issues"
"Source Code" = "https://github.com/Charlesshen2/img2excel"

[tool.setuptools]
packages = ["img2excel"]
